Use VirtualStudy service from public VS controller
Submitting dynamic virtual studies without explicitly defined sample IDs previously resulted in a NullPointerException.
This behavior was incorrect, as dynamic studies are designed to derive their sample IDs dynamically from filters, not require them upfront.
The issue revealed that the public VS controller was not
using the appropriate VirtualStudy service. This change aligns the controller
with the correct service to handle public dynamic virtual studies safely.
